refactor: no set default display
